Strictly enforce the node version required
This is helpful to avoid versions that will fail and with loud npm
install output, warnings can go unseen. With this change because of the
engine requirements specified upstream by the minecraft-protocol package
if someone attempts to npm install with an unsupported version of node
the process will abort with an error.